V0.1.1

+++ main.py
@@ -1,41 +1,49 @@
 """
 UselessOS
-By   the-cat1
-time 2022/7/16
+By      the-cat1
+Version 0.1.1
+Time    2022/7/17
 """
 
+import json
 import os
+import shutil
 import sys
 import time
 
-DRIVER = "Driver"
+VERSION = "0.1.1"
+VER_MAG = f"UselessOS(无用系统) {VERSION} [2022/7/17]，By the-cat1，Github https://github.com/the-cat1/UselessOS.git。"
 
+DRIVER = "Driver"
+HELP = os.path.abspath("HelpFiles\\")
+# noinspection SpellCheckingInspection
+HELP_FILE_SPLITEXT = ".uselessoshelp"
+
+helpDocs = {}
 directory = DRIVER  # 目录
 
 
+def helpInit():
+    """初始化Help系统"""
+
+    """获取帮助文件"""
+    helpFiles = os.listdir(HELP)
+    for file in range(len(helpFiles)):
+        if os.path.splitext(helpFiles[file])[1] != HELP_FILE_SPLITEXT:  # 后缀名不是HELP_FILE_SPLITEXT，即不是帮助文件
+            del helpFiles[file]
+
+    """录入帮助信息"""
+    for file in helpFiles:
+        with open(os.path.abspath(os.path.join(HELP, file)), encoding="utf-8") as f:
+            # noinspection PyBroadException
+            try:
+                helpJson = json.load(f)
+                helpDocs.update({helpJson["command"]: "\n".join(helpJson["helps"])})
+            except:
+                print(f"\033[33m\"{file}\"帮助文件存在问题！\033[0m")
+
+
 def cd(args: tuple):
-    """
-    cd [directory]
-
-    directory   要切换的目录。
-
-    将目录切换到指定的目录。
-    在不带参数时，将返回当前目录。
-
-
-    特殊目录：
-    .    当前目录
-    ..   上级目录
-
-
-    例如：
-        cd test
-    上面的例子将当前目录设置为"当前目录\\test"(当前目录存在"test"文件夹)。
-
-        cd
-    上面的例子将返回当前目录。
-    """
-
     global directory
 
     if len(args) == 0:
@@ -53,21 +61,6 @@ def cd(args: tuple):
 
 
 def mkdir(args: tuple):
-    """
-    mkdir (dir)
-
-    dir    要创建的目录。
-
-    创建单层目录。
-
-    注：mkdir命令只能创建单层目录，要创建多层目录，请用makedirs。
-
-
-    例子：
-        mkdir test
-    上面的例子在当前目录创建了一个"test"目录。
-    """
-
     if len(args) == 0:
         print("\033[31m\"mkdir\"命令需要一个参数！\033[0m\n")
     else:
@@ -80,19 +73,6 @@ def mkdir(args: tuple):
 
 
 def makedirs(args: tuple):
-    """
-    makedirs (dirs)
-
-    dirs   要创建的目录。
-
-    创建多层目录。
-
-
-    例子：
-        makedirs test1\\test2
-    上面的例子在当前目录创建了一个"test1"目录，又在"test1"目录中创建了"test2"目录。
-    """
-
     if len(args) == 0:
         print("\033[31m\"makedirs\"命令需要一个参数！\033[0m\n")
     else:
@@ -105,18 +85,6 @@ def makedirs(args: tuple):
 
 
 def newfile(args: tuple):
-    """
-    newfile (filename)
-
-    filename 文件名。
-
-    新建一个文件。
-
-    例子：
-        newfile test.txt
-    上面的例子在当前目录中创建了一个名为"test.txt"的文件。
-    """
-
     if len(args) == 0:
         print("\033[31m\"newfile\"命令需要一个参数！\033[0m\n")
     else:
@@ -131,19 +99,6 @@ def newfile(args: tuple):
 
 
 def type_(args: tuple):
-    """
-    type (filename)
-
-    filename 文件名。
-
-    输出某个文件的内容。
-
-
-    例子：
-        type test.txt
-    上面的例子将输出当前目录下的"test.txt"文件内容(当前目录下存在"test.txt")。
-    """
-
     if len(args) == 0:
         print("\033[31m\"type\"命令需要一个参数！\033[0m\n")
     else:
@@ -159,17 +114,6 @@ def type_(args: tuple):
 
 
 def dir_(args: tuple):
-    """
-    dir
-
-    返回当前目录下的所有文件和文件夹。
-
-
-    例子：
-        dir
-    上面的命令将返回当前目录下的所有文件和文件夹。
-    """
-
     if len(args) != 0:
         print("\033[31m\"dir\"没有参数！\033[0m\n")
         return
@@ -196,77 +140,43 @@ def dir_(args: tuple):
 
 
 def rm(args: tuple):
-    """
-    rm (filename)
-
-    filename 文件名。
-
-    删除某一文件。
-
-    注：rm只能删除文件，要删除文件夹，请用rmdir。
-
-
-    例子：
-        rm test.txt
-    上面的例子将会删除当前目录下的"test.txt"(当前目录存在"test.txt")。
-    """
-
     if len(args) == 0:
         print("\033[31m\"rm\"命令需要一个参数！\033[0m")
     else:
         fPath = os.path.abspath(os.path.join(directory, args[0]))
-        try:
-            os.remove(fPath)
-            print(f"{args[0]} 已删除。")
-        except OSError:
-            print("\033[31m删除失败！请检查目录是否正确！\033[0m")
-            print("\033[31m\"rmdir\"只能删除目录，要删除文件，请用\"rm\"\033[0m")
+        if os.path.isfile(fPath):
+            try:
+                os.remove(fPath)
+                print(f"{args[0]} 已删除。")
+            except OSError:
+                print("\033[31m删除失败！请检查目录是否正确！\033[0m")
+        else:
+            print("\033[31m\"rm\"只能删除文件，要删除文件夹，请用\"rmdir\"\033[0m")
     print("\n")
 
 
 def rmdir(args: tuple):
-    """
-    rmdir (dirname)
-
-    dirname 文件夹名。
-
-    删除某一文件夹(文件夹为空)。
-
-    注：rm只能删除文件，要删除文件夹，请用rmdir。
-
-
-    例子：
-        rmdir test
-    上面的例子将会删除当前目录下的"test"文件夹(当前目录存在"test"文件夹)。
-    """
-
     if len(args) == 0:
         print("\"rmdir\"命令需要一个参数！")
     else:
         dPath = os.path.abspath(os.path.join(directory, args[0]))
-        try:
-            os.rmdir(dPath)
-            print(f"{args[0]} 已删除。")
-        except OSError:
-            print("\033[31m删除失败！请检查目录是否正确！\033[0m")
+        if os.path.isdir(dPath):
+            # 删除所有文件
+            if "/a" in args:
+                shutil.rmtree(dPath)
+                print(f"{os.path.join(directory, args[0])} 及其子文件夹已删除。")
+            else:
+                try:
+                    os.rmdir(dPath)
+                    print(f"{os.path.join(directory, args[0])} 已删除。")
+                except OSError:
+                    print("\033[31m删除失败！请检查目录是否正确、为空！\033[0m")
+        else:
             print("\033[31m\"rmdir\"只能删除目录，要删除文件，请用\"rm\"\033[0m")
     print("\n")
 
 
 def rename(args: tuple):
-    """
-    rename (filename) (newName)
-
-    filename 文件名。
-    newName  新文件名。
-
-    重命名某一文件。
-
-    例子：
-        rename test.txt hello.txt
-    上面的例子会将当前目录下的"test.txt"重命名为"hello.txt"。
-    """
-
     if len(args) != 2:
         print("\033[31m\"rename\"命令需要2个参数！\033[0m\n")
     else:
@@ -278,21 +188,6 @@ def rename(args: tuple):
 
 
 def help_(args: tuple):
-    """
-    help [command]
-
-    command 命令
-
-    返回对应命令的帮助信息。
-
-    例子：
-        help
-    上面的例子会返回命令列表。
-
-        help cd
-    上面的例子会返回cd命令的帮助信息。
-    """
-
     if len(args) == 0:
         print("Help帮助：\n"
               "<cd>\t\t\t跳转到某个目录。\n"
@@ -306,42 +201,17 @@ def help_(args: tuple):
               "<rename>\t\t重命名文件。\n"
               "<rm>\t\t\t删除文件。\n"
               "<rmdir>\t\t\t删除文件夹。\n"
-              "<type>\t\t\t输出文件的内容。\n")
+              "<type>\t\t\t输出文件的内容。\n"
+              "<ver>\t\t\t输出UselessOS的版本信息。\n")
     else:
-        helpDocs = {
-            "cd": cd.__doc__,
-            "exit": exit_.__doc__,
-            "mkdir": mkdir.__doc__,
-            "makedirs": makedirs.__doc__,
-            "newfile": newfile.__doc__,
-            "type": type_.__doc__,
-            "dir": dir_.__doc__,
-            "ls": dir_.__doc__,
-            "rm": rm.__doc__,
-            "rmdir": rmdir.__doc__,
-            "help": help_.__doc__,
-            "rename": rename.__doc__,
-            "clear": clear.__doc__
-        }
         try:
             print(helpDocs[args[0]])
         except KeyError:
-            print("\033[31m指令不存在！\033[0m")
-        print("\n")
+            print("\033[31m指令不存在或缺少帮助文件！\033[0m")
+        print("")  # end="\n"
 
 
 def exit_(args: tuple):
-    """
-    exit
-
-    退出UselessOS。
-
-
-    例子：
-        exit
-    上面的例子会退出UselessOS。
-    """
-
     if len(args) != 0:
         print("\033[31m\"exit\"没有参数！\033[0m\n")
         return
@@ -350,17 +220,12 @@ def exit_(args: tuple):
 
 
 # noinspection PyUnusedLocal
+def ver(args: tuple):
+    print(VER_MAG)
+
+
+# noinspection PyUnusedLocal
 def clear(args: tuple):
-    """
-    clear
-
-    清除屏幕上的所有文字。
-
-
-    例子：
-        clear
-    上面的例子会清除屏幕上的所有文字。
-    """
     os.system("CLS")
 
 
@@ -378,7 +243,8 @@ commands = {
     "rmdir": rmdir,
     "help": help_,
     "rename": rename,
-    "clear": clear
+    "clear": clear,
+    "ver": ver
 }
 
 
@@ -386,7 +252,23 @@ def main():
     # 设置命令行
     os.system("CHCP 65001 && CLS")  # 确保能使用彩色字体
 
+    # 打印欢迎信息
     print("*****Welcome to UselessOS!*****")
+    print(VER_MAG)
+
+    """初始化"""
+    # 判断"Driver\"文件夹是否存在
+    if not os.path.isdir(os.path.abspath(DRIVER + "\\")):
+        print("\033[33m\"Driver\"文件夹不存在，已自动创建！\033[0m")
+        os.mkdir(os.path.abspath(DRIVER))
+
+    # 判断"Help\"文件夹是否存在
+    if not os.path.isdir(HELP):
+        print("\033[33m\"HELP\"文件夹不存在，已自动创建！\033[0m")
+        os.mkdir(HELP)
+
+    helpInit()
+
     while True:
         command = input(f"{directory}>")
         if command == "":  # 空指令
